+++ app/src/main/java/com/example/cicompanion/calendar/CalendarMode.kt
@@ -1,0 +1,7 @@
+package com.example.cicompanion.calendar
+
+enum class CalendarMode {
+    DAY,
+    WEEK,
+    MONTH
+}
