+++ app/src/main/java/com/example/cicompanion/calendar/model/CalendarSummary.kt
@@ -1,0 +1,7 @@
+package com.example.cicompanion.calendar.model
+
+data class CalendarSummary(
+    val id: String,
+    val title: String,
+    val isPrimary: Boolean
+)
