+++ app/src/main/java/com/example/cicompanion/calendar/model/EventsPage.kt
@@ -1,0 +1,6 @@
+package com.example.cicompanion.calendar.model
+
+data class EventsPage(
+    val items: List<CalendarEvent>,
+    val nextSyncToken: String?
+)
