--- app/src/main/java/com/example/sprint1homeui/calendar/CalendarMode.kt
@@ -1,7 +1,0 @@
-package com.example.sprint1homeui.calendar
-
-enum class CalendarMode {
-    DAY,
-    WEEK,
-    MONTH
-}
