--- app/src/main/java/com/example/sprint1homeui/calendar/model/CalendarSummary.kt
@@ -1,7 +1,0 @@
-package com.example.sprint1homeui.calendar.model
-
-data class CalendarSummary(
-    val id: String,
-    val title: String,
-    val isPrimary: Boolean
-)
