--- app/src/main/java/com/example/sprint1homeui/calendar/model/EventsPage.kt
@@ -1,6 +1,0 @@
-package com.example.sprint1homeui.calendar.model
-
-data class EventsPage(
-    val items: List<CalendarEvent>,
-    val nextSyncToken: String?
-)
