--- app/src/main/java/com/example/sprint1homeui/social/ProfileScreen.kt
@@ -1,179 +1,0 @@
-package com.example.sprint1homeui.social
-
-// unused imports are for the yet-to-be-implemented, fleshed-out top bar
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
-import com.example.sprint1homeui.ui.theme.Sprint1HomeUITheme
-import androidx.compose.material3.TopAppBar
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ProfileScreen(navController: NavHostController) {
-    Scaffold(
-        // topBar is a stub.
-        // for consistency, my (Noah) specific top bar elements are commented out,
-        // it may be useful to refer to this fleshed-out bar in the future
-        topBar = {
-            TopAppBar(
-                title = { Text("Profile", fontSize = 20.sp) }
-            )
-        }
-        /*
-            CenterAlignedTopAppBar(
-                title = {
-                    Text(
-                        text = "Profile",
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 20.sp
-                    )
-                },
-                navigationIcon = {
-                    IconButton(onClick = { /* TODO */ }) {
-                        Icon(
-                            painter = painterResource(id = R.drawable.ic_menu),
-                            contentDescription = "Menu",
-                            modifier = Modifier.size(24.dp)
-                        )
-                    }
-                },
-                actions = {
-                    IconButton(onClick = { /* TODO */ }) {
-                        Icon(
-                            painter = painterResource(id = R.drawable.ic_settings),
-                            contentDescription = "Settings",
-                            modifier = Modifier.size(24.dp)
-                        )
-                    }
-                },
-                colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
-                    containerColor = Color.White
-                )
-            )
-        }
-        */
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .padding(innerPadding)
-                .fillMaxSize()
-                .background(Color.White)
-        ) {
-            ProfileHeader(
-                modifier = Modifier
-                    .padding(horizontal = 24.dp, vertical = 32.dp)
-                    .fillMaxWidth()
-            )
-
-            // Gray content area
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(1f)
-                    .background(Color(0xFFE0E0E0)), // Light gray background
-                contentAlignment = Alignment.Center
-            ) {
-                // probably removable, but leaving for now
-                Button(onClick = { navController.navigate("home") }) {
-                    Text("Back to Home")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ProfileHeader(modifier: Modifier = Modifier) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Profile image placeholder
-        Box(
-            modifier = Modifier
-                .size(120.dp)
-                .clip(CircleShape)
-                .background(Color(0xFFE6E0F8)), // Light purple background
-            contentAlignment = Alignment.Center
-        ) {
-            // Silhouette
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center
-            ) {
-                Spacer(modifier = Modifier.height(10.dp))
-                Box(
-                    modifier = Modifier
-                        .size(40.dp)
-                        .clip(CircleShape)
-                        .background(Color(0xFF6750A4)) // Darker purple
-                )
-                Spacer(modifier = Modifier.height(4.dp))
-                Box(
-                    modifier = Modifier
-                        .size(80.dp, 40.dp)
-                        .clip(CircleShape)
-                        .background(Color(0xFF6750A4))
-                )
-            }
-        }
-
-        Spacer(modifier = Modifier.width(24.dp))
-
-        Column {
-            Text(
-                text = "User Name",
-                fontWeight = FontWeight.Bold,
-                fontSize = 24.sp,
-                color = Color.Black
-            )
-            Text(
-                text = "user@example.com",
-                color = Color.Gray,
-                fontSize = 16.sp
-            )
-            Spacer(modifier = Modifier.height(20.dp))
-            Text(
-                text = "Friends: 0",
-                fontWeight = FontWeight.Bold,
-                fontSize = 16.sp,
-                color = Color.Black
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ProfileScreenPreview() {
-    Sprint1HomeUITheme {
-        val navController = rememberNavController()
-        ProfileScreen(navController = navController)
-    }
-}
